dbms: development [#CONV-2944].

+++ dbms/include/DB/Columns/IColumn.h
@@ -100,7 +100,7 @@ public:
 	/** Размножить все значения столько раз, сколько прописано в offsets.
 	  * (i-е значение размножается в offsets[i] - offsets[i - 1] значений.)
 	  */
-	typedef UInt32 Offset_t;
+	typedef UInt64 Offset_t;
 	typedef std::vector<Offset_t> Offsets_t;
 	virtual void replicate(const Offsets_t & offsets) = 0;
